Likes, reads, auth and modal done
The like button is currently working completlely only if the user is signed in. The reads were shortened the code to be more efficient. The auth service is complete and sorking as it should along side of other components to show the data only if the user is signed in or is the user is the owner of the app. Additional the modal is currently working to show custom messages in the app.